[FEATURE] Replaced with use TimeUnit (#666)

+++ LaunchServer/src/main/java/pro/gravit/launchserver/components/AbstractLimiter.java
@@ -16,7 +16,7 @@ public abstract class AbstractLimiter<T> extends Component implements Reconfigur
     protected final transient Map<T, LimitEntry> map = new HashMap<>();
     private transient final Logger logger = LogManager.getLogger();
     public int rateLimit;
-    public int rateLimitMillis;
+    public long rateLimitMillis;
 
     @Override
     public Map<String, Command> getCommands() {
